Split catalog uploading into a separate object
