+++ handlers/catalog-service/src/main/scala/com/gu/catalogService/Handler.scala
@@ -1,14 +1,12 @@
 package com.gu.catalogService
 
-import com.amazonaws.services.s3.model.{PutObjectRequest, PutObjectResult}
-import com.gu.effects.{AwsS3, RawEffects}
+import com.gu.effects.RawEffects
 import com.gu.util.apigateway.LoadConfig
 import com.gu.util.reader.Types._
 import com.gu.util.zuora.{ZuoraRestConfig, ZuoraRestRequestMaker}
 import com.gu.util.{Logging, Stage}
-import java.io.{File, FileWriter}
 import okhttp3.{Request, Response}
-import play.api.libs.json.{JsValue, Json, Reads}
+import play.api.libs.json.{Json, Reads}
 import scala.util.Try
 
 object Handler extends Logging {
@@ -35,26 +33,6 @@ object Handler extends Logging {
     } yield ()
 
 
-  def uploadCatalogToS3(stage: Stage, catalog: JsValue): Try[PutObjectResult] = {
 
-    def jsonFile(catalog: JsValue): Try[File] = for {
-      file <- Try(new File("/tmp/catalog.json")) //Must use /tmp when running in a lambda
-      writer <- Try(new FileWriter(file))
-      _ <- Try(writer.write(catalog.toString()))
-      _ <- Try(writer.close())
-    } yield file
-
-    logger.info("Uploading catalog to S3...")
-
-    for {
-      catalogDotJson <- jsonFile(catalog)
-      putRequest = new PutObjectRequest(s"gu-zuora-catalog/${stage.value}", "catalog.json", catalogDotJson)
-      result <- AwsS3.putObject(putRequest)
-    } yield {
-      logger.info(s"Successfully uploaded file to S3: $result")
-      result
-    }
-
-  }
 
 }
